+++ src/features/SignUp/index.ts
@@ -1,0 +1,5 @@
+export { default as Input } from "./ui/Input";
+export { default as Terms } from "./ui/Terms";
+export { default as Title } from "./ui/Title";
+export { default as CheckInput } from "./ui/CheckInput";
+export { default as ConfirmButton } from "./ui/ConfirmButton";
