+++ src/features/SignUp/ui/ConfirmButton.tsx
@@ -1,0 +1,7 @@
+export default function ConfirmButton() {
+  return (
+    <button className="border-b-[1px] border-black pb-1 text-[16px]">
+      닉네임 중복 확인하고 가입 완료하기
+    </button>
+  );
+}
