+++ src/features/SignUp/ui/Title.tsx
@@ -1,0 +1,7 @@
+export default function Title() {
+  return (
+    <h3 className="h-[35px] text-[24px] font-black">
+      사용할 닉네임을 입력하세요.
+    </h3>
+  );
+}
