+++ src/widgets/SignUp/index.ts
@@ -1,0 +1,1 @@
+export { default as FormContainer } from "./ui/FormContainer";
